sw_engine: code refactoring.
rasterization functions are getting bigger.
split implementation vector-based and image-based for easier maintenance.
