+++ src/lib/sw_engine/meson.build
@@ -7,7 +7,8 @@ source_file = [
    'tvgSwImage.cpp',
    'tvgSwMath.cpp',
    'tvgSwRenderer.h',
-   'tvgSwRaster.cpp',
+   'tvgSwRasterVector.cpp',
+   'tvgSwRasterImage.cpp',
    'tvgSwRenderer.cpp',
    'tvgSwMemPool.cpp',
    'tvgSwRle.cpp',
